+++ usr/selftest/filter/median.c
@@ -14,10 +14,10 @@ static state median_testing(struct kshell_context *ctx, void *pdata)
 {
     struct median_state *test = pdata;
     unsigned int count;
-    short rawdata, filtered;
+    int32_t rawdata, filtered;
 
     for (count = 0; count < TEST_LOOP; ++count) {
-        rawdata = (short)prandom_value();
+        rawdata = (int32_t)prandom_value();
         filtered = median_update(test, rawdata);
         kshell_printf(ctx, "median filter test%02d: %6d -> %6d\n",
                       count, rawdata, filtered);
